refactor bill items
