+++ itemv2.js
@@ -49,26 +49,6 @@ const Option = {
   }
 };
 
-const SimpleOption = {
-  type: "object",
-  properties: {
-    name: {
-      type: "string",
-      title: "Display name of the option/variation/etc",
-      default: "",
-      examples: ["Soy Milk", "Extra Hot"]
-    },
-    amount: {
-      type: "integer",
-      title: "The cost of a single instance of this item.",
-      description:
-        "The cost is represented in the lowest denomination (eg cents). For discounts this should be negative.",
-      examples: [100, 0]
-    }
-  },
-  required: ["name", "amount"]
-};
-
 const BaseItem = {
   type: "object",
   properties: {
@@ -118,17 +98,6 @@ const BaseItem = {
   required: ["name", "quantity", "amount"]
 };
 
-const BillItem = {
-  type: "object",
-  properties: Object.assign({}, BaseItem.properties, {
-    changes: {
-      type: "array",
-      items: SimpleOption
-    }
-  }),
-  required: BaseItem.required
-};
-
 const OrderItem = {
   type: "object",
   properties: Object.assign({}, BaseItem.properties, {
@@ -140,4 +109,4 @@ const OrderItem = {
   required: BaseItem.required
 };
 
-module.exports = { Choice, Option, OrderItem, BillItem };
+module.exports = { BaseItem, Choice, Option, OrderItem };
